+++ CompanionPackages~/com.playserv.facebook-login/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## [0.5.1] - 2026-09-12
+
+- Aligned the package and internal PlayServ dependency versions with `0.5.1`.
+
 ## [0.5.0] - 2026-09-12
 
 - Aligned the package and internal PlayServ dependency versions with `0.5.0`.